+++ chrome/background.js
@@ -1,22 +1,7 @@
 /* background.js */
 
-import { extensionName, getOptions, logOptions } from './storage.js';
+import { extensionName, linkFormats, getOptions } from './storage.js';
 const debug = false;
-
-// Initialize notification icon and button tooltip
-const iconFilename = 'images/icon-48.png';
-const iconUrl = chrome.extension.getURL(iconFilename);
-getOptions().then(setTooltip);
-
-/*
-**  setTooltip: Set the extension button tooltip to show the
-**  extension name and the currently selected format option.
-*/
-function setTooltip (options) {
-  if (debug) logOptions('setTooltip', 'options', options);
-  let format = getCapitalizedFormat(options);
-  chrome.browserAction.setTitle({ title: `${extensionName}: ${format}` });
-}
 
 function getCapitalizedFormat (options) {
   switch (options.format) {
@@ -90,15 +75,8 @@ function processLinkData (data) {
   }
 
   function notifySuccess (options) {
-    let format = getCapitalizedFormat(options);
-    let message = `${format}-formatted link copied to clipboard.`;
-    let notificationOptions = {
-      type: 'basic',
-      iconUrl: iconUrl,
-      title: extensionName,
-      message: message
-    };
-    chrome.notifications.create(notificationOptions);
+    const format = linkFormats.get(options.format);
+    console.log(`Copied page link using ${format} format!`);
   }
 
   getOptions().then(function (options) {
@@ -108,25 +86,6 @@ function processLinkData (data) {
       notLastError();
     }
   });
-}
-
-/*
-**  copyPageLink: The handler for the browserAction.onClicked event and thus
-**  the main entry point to the extension.
-*/
-function copyPageLink (tab) {
-  // Security policy only allows us to inject the content script that
-  // accesses title and selection for pages loaded with http or https.
-  function checkUrlProtocol (tab) {
-    return (tab.url.indexOf('http:') === 0 || tab.url.indexOf('https:') === 0);
-  }
-
-  if (checkUrlProtocol(tab)) {
-    chrome.tabs.executeScript(null, { file: 'content.js' });
-  }
-  else {
-    processLinkData({ href: tab.url, title: '', selection: '' });
-  }
 }
 
 // Generic error handler
@@ -142,11 +101,6 @@ function notLastError () {
 
 function messageHandler (data, sender) {
   if (data.id === 'content') { processLinkData(data); }
-  if (data.id === 'tooltip') { setTooltip(data.options); }
 }
 
 chrome.runtime.onMessage.addListener(messageHandler);
-
-// Listen for toolbar button activation
-
-chrome.browserAction.onClicked.addListener(copyPageLink);