+++ firefox/background.js
@@ -1,22 +1,7 @@
 /* background.js */
 
-import { extensionName, getOptions, logOptions } from './storage.js';
+import { extensionName, linkFormats, getOptions } from './storage.js';
 const debug = false;
-
-// Initialize notification icon and button tooltip
-const iconFilename = 'images/icon-48.png';
-const iconUrl = browser.extension.getURL(iconFilename);
-getOptions().then(setTooltip);
-
-/*
-**  setTooltip: Set the extension button tooltip to show the
-**  extension name and the currently selected format option.
-*/
-function setTooltip (options) {
-  if (debug) logOptions('setTooltip', 'options', options);
-  let format = getCapitalizedFormat(options);
-  browser.browserAction.setTitle({ title: `${extensionName}: ${format}` });
-}
 
 function getCapitalizedFormat (options) {
   switch (options.format) {
@@ -90,42 +75,11 @@ function processLinkData (data) {
   }
 
   function notifySuccess (options) {
-    let format = getCapitalizedFormat(options);
-    let message = `${format}-formatted link copied to clipboard.`;
-    let notificationOptions = {
-      type: 'basic',
-      iconUrl: iconUrl,
-      title: extensionName,
-      message: message
-    };
-    return new Promise (function (resolve, reject) {
-      let promise = browser.notifications.create(notificationOptions);
-      promise.catch(
-        msg => { reject(new Error(`notifySuccess: ${msg}`)); }
-      );
-    });
+    const format = linkFormats.get(options.format);
+    console.log(`Copied page link using ${format} format!`);
   }
 
   getOptions().then(copyToClipboard).then(notifySuccess).catch(onError);
-}
-
-/*
-**  copyPageLink: The handler for the browserAction.onClicked event and thus
-**  the main entry point to the extension.
-*/
-function copyPageLink (tab) {
-  // Security policy only allows us to inject the content script that
-  // accesses title and selection for pages loaded with http or https.
-  function checkUrlProtocol (tab) {
-    return (tab.url.indexOf('http:') === 0 || tab.url.indexOf('https:') === 0);
-  }
-
-  if (checkUrlProtocol(tab)) {
-    browser.tabs.executeScript(null, { file: 'content.js' });
-  }
-  else {
-    processLinkData({ href: tab.url, title: '', selection: '' });
-  }
 }
 
 // Generic error handler
@@ -137,11 +91,6 @@ function onError (error) {
 
 function messageHandler (data, sender) {
   if (data.id === 'content') { processLinkData(data); }
-  if (data.id === 'tooltip') { setTooltip(data.options); }
 }
 
 browser.runtime.onMessage.addListener(messageHandler);
-
-// Listen for toolbar button activation
-
-browser.browserAction.onClicked.addListener(copyPageLink);